+++ packages/terra-consumer-nav/CHANGELOG.md
@@ -22,6 +22,7 @@ ChangeLog
 - Padding on nav items changed to be align with profile
 - Rendered translated texts using component SafeHtml.
 - Signout and signin links are now 'a' tags to utilize the href attribute.
+- Fixed logo text issue in Firefox
 
 ------------------